% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{akapit}
\alias{akapit}
\title{Drukowanie akapitow}
\usage{
akapit(tresc)
}
\arguments{
\item{tresc}{wektor ciągów znakowych z treścią kolejnych akalpitów.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja drukuje kod akapitów.
}
