% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_modele_ewd.R
\name{daj_naglowek_wykresu}
\alias{daj_naglowek_wykresu}
\title{Naglowek wykresu}
\usage{
daj_naglowek_wykresu(tekst, wykresStr = "Wykres")
}
\arguments{
\item{tekst}{ciąg znaków - opis tabeli}

\item{wykresStr}{opcjonalnie ciąg znaków z opisem typu obiektu, którego ma
to być nagłówek}
}
\value{
Funkcja zwraca kod Rmarkdown opisujący nagłówek wykresu
}
\description{
Funkcja generuje kod Rmarkdown opisujący nagłówek wykresu.
}

