% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drukuj_rysunki}
\alias{drukuj_rysunki}
\title{Drukowanie kodu latex dla tabeli i rysunków.}
\usage{
drukuj_rysunki(plikiRys, wymiar, znacznikiTab, tabelaSzablonTex,
  szerokosc = 1L)
}
\arguments{
\item{plikiRys}{wektor zawierający ścieżki do plików.}

\item{wymiar}{wymiar tabeli.}

\item{znacznikiTab}{lista z argumentami szablonu.}

\item{tabelaSzablonTex}{ścieżka do pliku z szablonem tabeli.}

\item{szerokosc}{szerokość tabeli z rysunkami. Jeden oznacza całą szerokość strony.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja drukuje kod latex dla szablinu tabeli z rysunkami zadanymi przez wektor ścieżek do plików.
}

