% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{drukuj_szablon}
\alias{drukuj_szablon}
\title{Drukowanie szablonu latex}
\usage{
drukuj_szablon(plikSzablonuTex, znaczniki)
}
\arguments{
\item{plikSzablonuTex}{ścieżka do szablonu}

\item{znaczniki}{lista definiująca znaczniki do zabawy}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Drukuje szablon z pliku oraz zamienia znaczniki na ciągi znakowe zdefiniowane
przez użytkownika. Jeśli w ramach pakietu został zdefiniowany plik
z szablonem o nazwie takiej, jak podana argumentem \code{plikSzablonuTex},
zostanie on wczytany z zasobów pakietu. W przeciwnym wypadku, argument
\code{plikSzablonuTex} zostanie potraktowany jako ścieżka do pliku
z szablonem względem aktywnego katalogu roboczego.
}

