% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drukuj_tabele}
\alias{drukuj_tabele}
\title{Tabela latex}
\usage{
drukuj_tabele(tabela, plikTabeliTex, znaczniki = NULL, ostatniWiersz = NULL,
  precyzja = 3, naglowek = TRUE, zamien_znaki = TRUE)
}
\arguments{
\item{tabela}{dane do zaprezentowania w formie tabeli}

\item{plikTabeliTex}{ścieżka pliku szablonu latex.}

\item{znaczniki}{lista argumentów, ktore występują w pliku szablonu tabeli.}

\item{ostatniWiersz}{wektor zawierający ostatni wiersza tabeli.}

\item{precyzja}{liczba miejsc po przecinku wyświetlanych dla tabeli}

\item{naglowek}{zmienna boolowska określająca, czy drukować nagłówek.}

\item{zamien_znaki}{zmienna określająca, czy zamieniać znaki specjalne latex'owe.}
}
\value{
Funkcja nie zwraca żadnej wartości.
}
\description{
Funkcja przekształca tabelę do kodu tex.
}

