% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{drukuj_wiersz_tabeli}
\alias{drukuj_wiersz_tabeli}
\title{Drukuj wiersz tabeli}
\usage{
drukuj_wiersz_tabeli(linia, wektor)
}
\arguments{
\item{linia}{linia z szablonu tabeli.}

\item{wektor}{wartości komórek danego wiersza tabeli.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja drukuje kod pojedynczej tabeli latex'owej.
}

