% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{inicjalizuj_numeracje}
\alias{inicjalizuj_numeracje}
\alias{initializuj_numeracje}
\title{Inicjalizacja numeracji}
\usage{
inicjalizuj_numeracje(nazwaZmiennejGlobalnej = "numeracjaTexZG",
  envir = .GlobalEnv)

initializuj_numeracje(nazwaZmiennejGlobalnej = "numeracjaTexZG",
  envir = .GlobalEnv)
}
\arguments{
\item{nazwaZmiennejGlobalnej}{nazwa zmiennej globalnej}

\item{envir}{środowisko, gdzie znajduje się zmienna o nazwie
nazwaZmiennejGlobalnej}
}
\value{
Funkcja zwraca obiekt klasy 'NumeracjaTex'.
}
\description{
Funkcja inicjalizuje numeracje. Używając numeracji nie powinno się
zmieniać zmiennej globalnej określonej parametrem.
}
\section{Functions}{
\itemize{
\item \code{initializuj_numeracje}: alias (pierwotna nazwa nadana przez Grześka)
}}

