% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initializuj_numeracje}
\alias{initializuj_numeracje}
\title{Initializacja numeracji}
\usage{
initializuj_numeracje(nazwaZmiennejGlobalnej = "numeracjaTexZG",
  envir = .GlobalEnv)
}
\arguments{
\item{nazwaZmiennejGlobalnej}{nazwa zmiennej globalnej.}

\item{envir}{środowisko, gdzie znajduje się zmienna o nazwie nazwaZmiennejGlobalnej.}
}
\value{
Funkcja zwraca obiekt klasy 'NumeracjaTex'.
}
\description{
Funkcja initializuje numeracje. Używając numeracji nie powinno się
zmieniać zmiennej globalnej określonej parametrem.
}

