% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_modele_ewd.R
\name{kable_table}
\alias{kable_table}
\title{Drukowanie tabeli}
\usage{
kable_table(tab, naglowek = NULL, format.args = list(decimal.mark =
  ",", big.mark = " "))
}
\arguments{
\item{tab}{ramka danych lub}

\item{naglowek}{treść nagłówka}

\item{format.args}{opcjonalnie list dodatkowych argumentów opisujących
formatowanie liczb w tabeli (p. \code{\link[knitr]{kable}})}
}
\value{
Funkcja zwraca kod Rmarkdown, zwrócony przez funkcję
\code{\link[knitr]{kable}}.
}
\description{
Funkcja przygotowuje nieco przekazaną jej macierz lub ramkę danych,
a następnie zamienia ją na kod Rmarkdown przy pomocy funkcji.
\code{\link[knitr]{kable}}
}
