% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{lista_numerowana}
\alias{lista_numerowana}
\title{Lista numerowana}
\usage{
lista_numerowana(elementyListy)
}
\arguments{
\item{elementyListy}{ciągi znakowe będące elementami listy}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja drukuje kod listy numerowanej.
}

