% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lista_punktowana}
\alias{lista_punktowana}
\title{Lista punktowana}
\usage{
lista_punktowana(elementyListy)
}
\arguments{
\item{elementyListy}{ciągi znakowe będące elementami listy.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja drukuje kod listy punktowanej
}

