% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{numeracja}
\alias{numeracja}
\title{Numeracja}
\usage{
numeracja(nazwa, tylkoNumer = FALSE,
  nazwaZmiennejGlobalnej = "numeracjaTexZG", envir = .GlobalEnv)
}
\arguments{
\item{nazwa}{nazwa numerowanego obiektu, np.: 'Tabela'.}

\item{tylkoNumer}{jeżeli TRUE zwróci numer obiektu}

\item{nazwaZmiennejGlobalnej}{nazwa zmiennej globalnej.}

\item{envir}{środowisko, gdzie znajduje się zmienna o nazwie nazwaZmiennejGlobalnej.}
}
\value{
Funkcja zwraca ciąg znaków.
}
\description{
Funkcja dokleja do parametru nazwa numer, który oznacza kolejność wywołania funcji z tym parametrem.
Używając numeracji nie powinno się zmieniać zmiennej globalnej 'numeracjaTex'.
}

