% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_modele_ewd.R
\name{pander_table}
\alias{pander_table}
\title{Drukowanie tabeli}
\usage{
pander_table(tab, ...)
}
\arguments{
\item{tab}{ramka danych lub macierz}

\item{...}{opcjonalnie dodatkowe argumenty, które mają zostać przekazane
funkcji \code{\link[pander]{pander}}}
}
\value{
Funkcja zwraca kod Rmarkdown, zwrócony przez funkcję
\code{\link[pander]{pander}}
}
\description{
Funkcja przygotowuje nieco przekazaną jej macierz lub ramkę danych,
a następnie zamienia ją na kod Rmarkdown przy pomocy funkcji.
\code{\link[pander]{pander}}
}

