% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_modele_ewd.R
\name{parsuj_plik_porownanie}
\alias{parsuj_plik_porownanie}
\title{Wczytywanie wlasnosci modeli}
\usage{
parsuj_plik_porownanie(plik, pierwszyWiersz = 1, encoding = "windows-1250")
}
\arguments{
\item{plik}{ciąg znaków - nazwa pliku}

\item{pierwszyWiersz}{liczba całkowita - od którego wiersza zacząć
przetwarzanie pliku}

\item{encoding}{ciąg znaków określający kodowanie pliku}
}
\value{
Funkcja zwraca listę ramek danych.
}
\description{
Funkcja wczytuje plik opisujący własności modeli EWD o różnych stopniach
wielomianu opisującego zależność pomiędzy wynikami "na wejściu" a wynikami
"na wyjściu".
}

