% Generated by roxygen2 (4.0.2): do not edit by hand
\name{podrozdzial}
\alias{podrozdzial}
\title{Drukowanie podrozdziału}
\usage{
podrozdzial(tytul)
}
\arguments{
\item{tytul}{tytuł podrozdziału}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Fukcja drukujaca kod latex z tytułem podrozdziału
}

