% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/raporty.R
\name{raport_laczenie}
\alias{raport_laczenie}
\title{Raport łączenie}
\usage{
raport_laczenie(tabelaPolaczen, nazwaSkalowania, tabelaSzablonTex,
  znacznikiTabeli, tekstAkapitu, tabelaSzablonRysunkiTex,
  znacznikiTabeliRysunkow, ktoreRys, folderWykresow, szerokosc = 1)
}
\arguments{
\item{tabelaPolaczen}{tabela do wyświetlenia.}

\item{nazwaSkalowania}{nazwa skalowania, która służy do stworzenia nazw plików wejściowych.}

\item{tabelaSzablonTex}{ścieżka do szablonu tabeli latex.}

\item{znacznikiTabeli}{lista ze znacznikami szablonu tabeli.}

\item{tekstAkapitu}{tekst akapitu pomiędy tabelą i wykresami}

\item{tabelaSzablonRysunkiTex}{ścieżka do szablonu tabeli, w której znajdą się wykresy}

\item{znacznikiTabeliRysunkow}{lista ze znacznikami szablonu tabeli rysunków.}

\item{ktoreRys}{wektor indeksów określające, które wykresy należy wyświetlić.}

\item{folderWykresow}{ścieżka do folderów z wykresami.}

\item{szerokosc}{ułamek określający jaką część strony powinna zajmować tabela z rysunkami.}
}
\value{
Funkcja nie zwraca żadnej wartości.
}
\description{
Funkcja drukuje kod latex raportu z łączenia.
}

