% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renderuj_raport_Rmd.R
\name{renderuj_raport_Rmd}
\alias{renderuj_raport_Rmd}
\title{Funkcja renderujaca plik szablonu RMD}
\usage{
renderuj_raport_Rmd(inputFile, clean = FALSE, pdf = TRUE, tex = FALSE,
  envir = .GlobalEnv)
}
\arguments{
\item{inputFile}{nazwa pliku rmd z szablonem raportu.}

\item{clean}{jeżeli ustawiony na TRUE to usuwane są wszystkie pliki robocze.
(UWAGA: Mogą zostać usunięte inne pliki, których nazwy na początku zawierają nazwę szablonu).}

\item{pdf}{jeżeli ustawiony na TRUE to generuje plik pdf.}

\item{tex}{jeżeli ustawiony na TRUE to generuje plik tex.}

\item{envir}{środowisko renderowania szablonu rmd. Powinno zawierać wszystkie stałe globalne, które używane są w szablonie}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja służy do renderowania plików rmd na formaty: pdfi tex.
}

