% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_modele_ewd.R
\name{ustaw_naglowek_tabeli}
\alias{ustaw_naglowek_tabeli}
\title{Naglowek tabeli}
\usage{
ustaw_naglowek_tabeli(tekst, tabelaStr = "Tabela")
}
\arguments{
\item{tekst}{ciąg znaków - opis tabeli}

\item{tabelaStr}{opcjonalnie ciąg znaków z opisem typu obiektu, którego ma
to być nagłówek}
}
\value{
Funkcja zwraca kod Rmarkdown, zwrócony przez funkcję
\code{\link[pander]{set.caption}}
}
\description{
Funkcja generuje kod Rmarkdown opisujący nagłówek tabeli.
}

