% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{zamien_znaki}
\alias{zamien_znaki}
\title{Zamien znaki}
\usage{
zamien_znaki(nazwa, strOrg, strN)
}
\arguments{
\item{nazwa}{ciąg znaków do zmiany}

\item{strOrg}{wektor ciągów znaków do zastąpienia.}

\item{strN}{wektor ciągów znaków zastępujące odpowiednie elementy z
\code{strOrg}.}
}
\value{
Przekształcony ciąg znaków.
}
\description{
Funkcja przekształca ciąg znaków zastępując podciągi (zadane argumentem
\code{strOrg}) na odpowiadające im podciągi zadane argumentem
\code{znakiNowe}.
}
