% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{zamien_znaki_tex}
\alias{zamien_znaki_tex}
\title{Zamień znaki specjalne.}
\usage{
zamien_znaki_tex(nazwa)
}
\arguments{
\item{nazwa}{ciąg znaków do zmiany}
}
\value{
Przekształcony ciąg znaków.
}
\description{
Funkcja przekształca ciąg znaków zastępując wybrane znak specjalne na kod tex, który pozwala
na wyświetlenie ich w dokumencie tex.
}

