% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{drukuj_szablon}
\alias{drukuj_szablon}
\title{Drukowanie szablonu latex}
\usage{
drukuj_szablon(plikSzablonuTex, znaczniki)
}
\arguments{
\item{plikSzablonuTex}{ścieżka do szablonu.}

\item{znaczniki}{lista definiująca znaczniki do zabawy.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Drukuje szablon z pliku oraz zamienia znaczniki na ciągi znakowe zdefiniowane przez użytkownika.
}

