% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{drukuj_wiersz_tex}
\alias{drukuj_wiersz_tex}
\title{Drukuj wiersz}
\usage{
drukuj_wiersz_tex(linia, znaczniki = NULL)
}
\arguments{
\item{linia}{linia kodu latex.}

\item{znaczniki}{lista zdefiniowanych znaczników.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja drukuje wiersz szablonu uwzględniając znacznki.
}
