% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knit_odra.R
\name{knit_odra}
\alias{knit_odra}
\title{Knit Odra}
\usage{
knit_odra(raport, folder)
}
\arguments{
\item{raport}{nazwa pliku knitr do wygenerowania raportu.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja służąca do budowania raportów za pomocą funkcji \link{knit} na serwerze Odra.
Wprowadzona w celu generowania raportów z poprawnymi polskimi znakami.
}

