% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knit_odra}
\alias{knit_odra}
\title{Knit Odra}
\usage{
knit_odra(file)
}
\arguments{
\item{file}{ścieżka względna do plik z rozszerzeniem 'Rnw'.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja służąca do budowania raportów za pomocą funkcji \link{knit} na serwerze Odra.
Wprowadzona w celu generowania raportów z poprawnymi polskimi znakami.
}

