% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numeracja}
\alias{numeracja}
\title{Numeracja}
\usage{
numeracja(nazwa, nazwaZmiennejGlobalnej = "numeracjaTexZG",
  envir = .GlobalEnv)
}
\arguments{
\item{nazwa}{nazwa numerowanego obiektu, np.: 'Tabela'.}

\item{nazwaZmiennejGlobalnej}{nazwa zmiennej globalnej.}

\item{envir}{środowisko, gdzie znajduje się zmienna o nazwie nazwaZmiennejGlobalnej.}
}
\value{
Funkcja zwraca ciąg znaków.
}
\description{
Funkcja dokleja do parametru nazwa numer, który oznacza kolejność wywołania funcji z tym parametrem.
Używając numeracji nie powinno się zmieniać zmiennej globalnej 'numeracjaTex'.
}

