% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{podrozdzial}
\alias{podrozdzial}
\title{Drukowanie podrozdzialu}
\usage{
podrozdzial(tytul)
}
\arguments{
\item{tytul}{tytuł podrozdziału}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Fukcja drukujaca kod latex z tytułem podrozdziału
}
