% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderuj_raport_Rmd.R
\name{renderuj_raport_Rmd}
\alias{renderuj_raport_Rmd}
\title{Funkcja renderujaca plik szablonu RMD}
\usage{
renderuj_raport_Rmd(szablon, nazwaPliku, obiekty = list(), clean = TRUE)
}
\arguments{
\item{szablon}{nazwa pliku z szablonem raportu w formacie .Rmd}

\item{nazwaPliku}{nazwa pliku wynikowego (razem z rozszerzeniem określającym
format)}

\item{obiekty}{lista obiektów, z których korzysta szablon do wygenerowania
raportu}

\item{clean}{wartość logiczna - czy usunąć wszystkie pliki robocze generowane
przez pander przy tworzeniu pliku wynikowego?}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja służy do renderowania raportów na podstawie szablonów zapisanych
w plikach .Rmd. Jeśli w ramach pakietu został zdefiniowany plik
z szablonem o nazwie takiej, jak podana argumentem \code{szablon}, zostanie
on wczytany z zasobów pakietu. W przeciwnym wypadku, argument \code{szablon}
zostanie potraktowany jako ścieżka do pliku z szablonem względem aktywnego
katalogu roboczego.
}

