% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{rozdzial}
\alias{rozdzial}
\title{Drukowanie tytułu rozdziału}
\usage{
rozdzial(tytul)
}
\arguments{
\item{tytul}{tytuł rozdziału.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja drukujaca kod latex z tytułem rozdziału.
}

