% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{zamien_znaczniki}
\alias{zamien_znaczniki}
\title{Drukuj wiersz}
\usage{
zamien_znaczniki(nazwa, znaczniki)
}
\arguments{
\item{nazwa}{ciąg znaków}

\item{znaczniki}{lista znaczników i przypisane im ciągi znakowe}
}
\value{
Funkcja zwraca ciąg znaków z zamienionymi znacznikami na zdefiniowane
w parametrze znaczniki ciągi znaków.
}
\description{
Funkcja zamienia znaczniki na zdefiniowane ciągi znaków.
}

