% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zamien_znaki}
\alias{zamien_znaki}
\title{Zamień znaki.}
\usage{
zamien_znaki(nazwa, strOrg, strN)
}
\arguments{
\item{nazwa}{ciąg znaków do zmiany}

\item{strOrg}{wektor ciągów znaków do zastąpienia.}

\item{strN}{wektor ciągów znaków zastępujące odpowiednie elementy z strOrg.}
}
\value{
Przekształcony ciąg znaków.
}
\description{
Funkcja przekształca ciąg znaków zastępując podciągi (zadane parametrem znakiOrg)
na odpowiadające im podciągi zadane parametrem znakiNowe.
}

